Extract parse modules in seperate file for better readiablity
