+++ parser/module.go
@@ -1,0 +1,33 @@
+package parser
+
+import (
+	"os"
+	"path/filepath"
+	"strings"
+)
+
+func (p *parser) parseModule() error {
+	walker := func(path string, info os.FileInfo, err error) error {
+		if info != nil && info.IsDir() {
+			if strings.HasPrefix(strings.Trim(strings.TrimPrefix(path, p.ModulePath), "/"), ".git") {
+				return nil
+			}
+			fns, err := filepath.Glob(filepath.Join(path, "*.go"))
+			if len(fns) == 0 || err != nil {
+				return nil
+			}
+			// p.debug(path)
+			name := filepath.Join(p.ModuleName, strings.TrimPrefix(path, p.ModulePath))
+			name = filepath.ToSlash(name)
+			p.KnownPkgs = append(p.KnownPkgs, pkg{
+				Name: name,
+				Path: path,
+			})
+			p.KnownNamePkg[name] = &p.KnownPkgs[len(p.KnownPkgs)-1]
+			p.KnownPathPkg[path] = &p.KnownPkgs[len(p.KnownPkgs)-1]
+		}
+		return nil
+	}
+	filepath.Walk(p.ModulePath, walker)
+	return nil
+}
